+++ philo/srcs/print.c
@@ -1,21 +1,21 @@
 #include "philo.h"
 
-static void	print_status(int status, long ms_time, t_philo *philo)
+static void	print_status(int status, long long ms_time, t_philo *philo)
 {
 	if (status == FORK)
-		printf("%ld %d has taken a fork\n", ms_time, philo->id);
+		printf("%lld %d has taken a fork\n", ms_time, philo->id);
 	else if (status == EAT)
 	{
-		printf("%ld %d is eating\n", ms_time, philo->id);
+		printf("%lld %d is eating\n", ms_time, philo->id);
 		check_count_eat(philo->info, philo);
 	}
 	else if (status == SLEEP)
-		printf("%ld %d is sleeping\n", ms_time, philo->id);
+		printf("%lld %d is sleeping\n", ms_time, philo->id);
 	else if (status == THINK)
-		printf("%ld %d is thinking\n", ms_time, philo->id);
+		printf("%lld %d is thinking\n", ms_time, philo->id);
 	else if (status == DIE)
 	{
-		printf("%ld %d died\n", ms_time, philo->id);
+		printf("%lld %d died\n", ms_time, philo->id);
 		switch_flag_to_fin();
 	}
 }
@@ -32,7 +32,7 @@ static int	check_fin(void)
 
 int	print_status_and_check_fin(t_philo *philo, int status)
 {
-	long	ms_time;
+	long long	ms_time;
 
 	pthread_mutex_lock(&g_print);
 	if (check_fin() == FINISH)
